add validation decorator to ensure file and folder lists are not empty before Perforce operations; move utility functions to new utils.py module
